// FRONTEND/src/types/appData.ts
interface LocationData {
  west_bound_longitude: `${number}` | null;
  east_bound_longitude: `${number}` | null;
  north_bound_longitude: `${number}` | null;
  south_bound_longitude: `${number}` | null;
  mean_latitude: `${number}` | null;
  mean_longitude: `${number}` | null;
}

export interface IDataset {
  id: IDatasetID;
  location_data: LocationData;
  doi: IDoi;
  dataset_publication_date?: IPublicationDate;
  temporal_coverage: TemporalCoverage | null;
  authors?: IAuthor[];
  providers?: IProvider[];
  dataset_title: ITitle;
  organization: "PANGAEA" | "Earth Observatory";
}

//Dataset attributes
export type IDatasetID =
  | `Dataset/${IDatasetKEY}`
  | `STACCollection/${IDatasetKEY}`;
export type IDatasetKEY = number | string;
export type IAbstract = string | null;
export type ITitle = string | null;
export type IAuthor = string | null;
export type IProvider = string | null;
export type IDoi = string | null;
export type IPublicationDate = string | null;

export interface IKeywordData {
  keyword: string;
  count: number;
  dataset_id: IDatasetID[];
}

export interface IGeoData {
  id: IDatasetID;
  coordinates: [
    lat: LocationData["mean_latitude"],
    lon: LocationData["mean_longitude"]
  ];
  groupId: "dataset" | "staccollection" | "active";
}

export interface ITransformedTimeData {
  start: string;
  end: string;
  dataset_title: ITitle;
}
export interface ITimeData {
  date: number | string;
  value: number;
}

export interface AuthorData {
  author: string;
  datasets: IDatasetID[];
}

export interface ChordData {
  from: string;
  to: string;
  value: number;
}

export interface IPointHoverHandler {
  (lat: string, lon: string): void;
}